Добавить поддержку байтов в файлах, рефактор транспортов и метод read_message
