+++ redocs/source/index.rst
@@ -5,6 +5,7 @@
 .. image:: _static/logo.svg
    :align: center
    :width: 320px
+
 PyMax
 =====
 
@@ -34,6 +35,7 @@ PyMax
    decorators
    examples
    guides
+   release_notes
 
 .. rubric:: Особенности
 
@@ -41,6 +43,10 @@ PyMax
 - Отправка / редактирование / удаление сообщений
 - Управление чатами, каналами и диалогами
 - Получение истории сообщений
+- Загрузка фотографий профиля
+- Разрешение групп по ссылке
+- Поддержка контактов в сообщениях
+- Управление списком контактов
 
 ---
