+++ redocs/source/release_notes.rst
@@ -1,0 +1,61 @@
+Release Notes v1.2.3
+====================
+
+Новые функции
+-------------
+
+**Загрузка фотографий профиля**
+    Профиль теперь может быть обновлен с загрузкой новой фотографии через метод ``change_profile()``.
+
+**Разрешение групп по ссылке**
+    Группы теперь могут быть разрешены (получены) прямо по их ссылке через метод ``resolve_group_by_link()``.
+
+**Поддержка контактов в сообщениях**
+    Сообщения теперь поддерживают вложения типа контакта с информацией о контакте (ContactAttach).
+
+**Список контактов клиента**
+    Клиент теперь ведет список всех контактов пользователя через свойство ``client.contacts``.
+
+Новые методы
+------------
+
+MaxClient.resolve_group_by_link(link: str) -> Chat | None
+    Разрешает группу по ссылке. Возвращает объект чата группы или None, если не найдено.
+
+MaxClient.change_profile(first_name, last_name, description, photo)
+    Изменяет информацию профиля текущего пользователя, включая загрузку новой фотографии.
+
+Новые типы
+----------
+
+ContactAttach
+    Представляет контакт в сообщении. Содержит информацию о контакте (ID, имя, фамилия, фото).
+
+Измененные типы
+---------------
+
+Message
+    Теперь поддерживает вложения типа ContactAttach в список attaches.
+
+Names
+    Улучшен для работы с различными форматами имен пользователя.
+
+StickerAttach
+    Улучшено представление стикеров в сообщениях.
+
+Photo
+    Улучшено для работы с фотографиями профиля.
+
+AttachType
+    Добавлено значение CONTACT для контактов.
+
+Новые параметры
+---------------
+
+MaxClient.contacts: list[User]
+    Список контактов текущего пользователя.
+
+Версия
+------
+
+**1.2.3** - выпущена 24 декабря 2025 г.
